+++ Singer.API/ClientApp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "Singer",
-  "version": "0.1.0-issue-AB-12-view-care-users.120",
+  "version": "0.1.0-issue-AB-12-view-care-users.126",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
